49 dto records (#94)
* TimeSeriesPointDto als AggregatePointDto  ins Shared-Projekt verschoben

* Dto von AccountsController in Shared verschoben

* Dtos von AdminIpBlocksController verschoben

* AdminUsersController

* AggregatesController

* AttachmentsController

* AuthController

* BackgroundTasksController

* BackupsController

* ContactCategoriesController

* ContactReportsController

* ContactsController

* DividendsReportsController

* HomeKpisController

* MetaHolidayProvidersController

* PostingsController

* PostingsExportController

* ReportAggregatesController

* ReportFavoritesController

* SavingsPlanCategoriesController

* SavingsPlanReportsController

* SavingsPlansController

* SecuritiesController

* SecurityBackfillController

* SecurityCategoriesController

* SecurityPricesController

* SecurityReportsController

* StatementDraftsController

* UserImportSplitSettingsController

* UserNotificationSettingsController

* UserProfileSettingsController

* UsersController

* Nutzung von PostingKind anstelle von int

* Weiterleitung bei fehlender Anmeldung auf Bankkontoseite korrigiert

* Xml-Kommentare für Shared-projekt

* Dtos in Unterordner gruppiert

* Nur ein Typ pro Datei in Shared

* Weitere Dtos verschoben

* BackupDto erfassen

* Verwendung der Dtos angepasst

* Verwendung der Dtos angepasst

* Controllerendpunkte thematisch zusammengefasst

* Controller kommentiert

* Erste ViewModel und Pages auf ApiClient umgestellt.

* AuthController im ApiClient implementiert

* Integrationstests für ApiClient + Migrationen korrigiert + UserLockout unterstzützt

* Aufrufe für BackroundTaskController im ApiClient implementiert + appSettings-Einstellung für Deaktivierung von Hintergrundaufgaben.

* appSettings aktualisiert

* Aufrufe des BackupController im ApiClient implementiert

* Aufrufe des ContactCategoryController im ApiClient implementiert

* Aufrufe des ContactsController im ApiClient implementiert

* Aufrufe des HomeKpisController im ApiClient implementiert

* Aufrufe des MetaHolidaysControllerim ApiClient implementiert

* Aufrufe für NotificationsController in ApiClient implementiert

* Aufrufe für PostingController im ApiClient

* ReportsController

* SavingsPlanCategories

* SavingPlanController

* securityController

* securityCategory

* Statement Drafts

* Das Prinzip von "Regsistrierung nur einmal" in der UI umgesetzt.

* UsersController

* UserSettingsController

* Letzte Aufrufe umgesteltl auf ApiClient + IHttpClientFactory in ViewModels und Pages entfernen + Globalusings mit allen Dto-Pfaden

+++ FinanceManager.Tests/Reports/ContactPostingsValutaDateTests.cs
@@ -1,19 +1,10 @@
-using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using FinanceManager.Application.Reports;
-using FinanceManager.Domain;
 using FinanceManager.Domain.Contacts;
+using FinanceManager.Domain.Postings;
 using FinanceManager.Infrastructure;
+using FinanceManager.Infrastructure.Aggregates;
 using FinanceManager.Infrastructure.Reports;
 using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
-using Xunit;
-using FinanceManager.Shared.Dtos;
-using FinanceManager.Domain.Reports;
-using FinanceManager.Domain.Postings;
-using FinanceManager.Infrastructure.Aggregates;
 
 namespace FinanceManager.Tests.Reports;
 
@@ -100,7 +91,7 @@ public sealed class ContactPostingsValutaDateTests
         var analysis = new DateTime(year, 2, 1);
         var query = new ReportAggregationQuery(
             OwnerUserId: user.Id,
-            PostingKind: (int)PostingKind.Contact,
+            PostingKind: PostingKind.Contact,
             Interval: ReportInterval.Month,
             Take: 12,
             IncludeCategory: false,
@@ -197,7 +188,7 @@ public sealed class ContactPostingsValutaDateTests
         var analysis = new DateTime(year, 2, 1);
         var query = new ReportAggregationQuery(
             OwnerUserId: user.Id,
-            PostingKind: (int)PostingKind.Contact,
+            PostingKind: PostingKind.Contact,
             Interval: ReportInterval.Month,
             Take: 12,
             IncludeCategory: false,
