+++ FinanceManager.Tests/Web/PostingsQueryServiceAdditionalTests.cs
@@ -1,18 +1,12 @@
-using FinanceManager.Web.Services;
-using FinanceManager.Infrastructure;
-using FinanceManager.Domain;
 using FinanceManager.Domain.Accounts;
 using FinanceManager.Domain.Contacts;
 using FinanceManager.Domain.Postings;
 using FinanceManager.Domain.Savings;
 using FinanceManager.Domain.Securities;
-using FinanceManager.Shared.Dtos;
+using FinanceManager.Infrastructure;
+using FinanceManager.Web.Services;
 using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Xunit;
 
 namespace FinanceManager.Tests.Web
 {
@@ -59,7 +53,7 @@ namespace FinanceManager.Tests.Web
             Assert.NotNull(res);
             Assert.Single(res);
             var dto = res[0];
-            Assert.Equal((int)PostingKind.Bank, dto.Kind);
+            Assert.Equal(PostingKind.Bank, dto.Kind);
             Assert.Equal(acc.Id, dto.BankPostingAccountId);
             Assert.Equal(sym, dto.BankPostingAccountSymbolAttachmentId);
         }
@@ -81,7 +75,7 @@ namespace FinanceManager.Tests.Web
             var res = await svc.GetSavingsPlanPostingsAsync(plan.Id, 0, 50, null, null, null, owner);
             Assert.NotNull(res);
             Assert.Single(res);
-            Assert.Equal((int)PostingKind.SavingsPlan, res[0].Kind);
+            Assert.Equal(PostingKind.SavingsPlan, res[0].Kind);
             Assert.Equal(plan.Id, res[0].SavingsPlanId);
         }
 
@@ -102,7 +96,7 @@ namespace FinanceManager.Tests.Web
             var res = await svc.GetSecurityPostingsAsync(sec.Id, 0, 50, null, null, owner);
             Assert.NotNull(res);
             Assert.Single(res);
-            Assert.Equal((int)PostingKind.Security, res[0].Kind);
+            Assert.Equal(PostingKind.Security, res[0].Kind);
             Assert.Equal(sec.Id, res[0].SecurityId);
         }
